// apps/api/src/remotion/components/ProgressBar.d.ts
import React from 'react';
interface ProgressBarProps {
    progress: number;
    width?: number;
    height?: number;
    backgroundColor?: string;
    fillColor?: string;
    animated?: boolean;
    label?: string;
    showPercentage?: boolean;
}
export declare const ProgressBar: React.FC<ProgressBarProps>;
export {};
//# sourceMappingURL=ProgressBar.d.ts.map